+++ packages/core/src/components/graphs/gauge.ts
@@ -330,7 +330,10 @@ export class Gauge extends Component {
 		// Draw the arrow with status
 		const status = Tools.getProperty(options, "gauge", "status");
 		DOMUtils.appendOrSelect(deltaArrow, "polygon.gauge-delta-arrow")
-			.attr("class", status !== null ? `gauge-delta-arrow status--${status}` : "")
+			.attr(
+				"class",
+				status !== null ? `gauge-delta-arrow status--${status}` : ""
+			)
 			.attr("fill", () => (status === null ? "currentColor" : null))
 			.attr("points", self.getArrow(delta));