+++ packages/core/src/services/zoom.ts
@@ -57,6 +57,7 @@ export class Zoom extends Service {
 			"top",
 			"data"
 		);
+
 		// if user already defines zoom bar data, use it
 		if (definedZoomBarData && definedZoomBarData.length > 1) {
 			zoomBarData = definedZoomBarData;
@@ -64,10 +65,13 @@ export class Zoom extends Service {
 			// use displayData if not defined
 			zoomBarData = this.model.getDisplayData();
 		}
+
 		// get all dates (Number) in displayData
 		let allDates = [];
 		zoomBarData.forEach((data) => {
-			allDates = allDates.concat(Number(data[domainIdentifier]));
+			allDates = allDates.concat(
+				new Date(data[domainIdentifier]).getTime()
+			);
 		});
 		allDates = Tools.removeArrayDuplicates(allDates).sort();
 		// Go through all date values
@@ -77,7 +81,7 @@ export class Zoom extends Service {
 			const datum = {};
 
 			zoomBarData.forEach((data) => {
-				if (Number(data[domainIdentifier]) === date) {
+				if (new Date(data[domainIdentifier]).getTime() === date) {
 					sum += data[rangeIdentifier];
 				}
 			});
@@ -101,13 +105,14 @@ export class Zoom extends Service {
 		);
 	}
 
-	handleDomainChange(newDomain) {
+	handleDomainChange(newDomain, configs = { dispatchEvent: true }) {
 		this.model.set({ zoomDomain: newDomain }, { animate: false });
-		this.services.events.dispatchEvent(Events.ZoomDomain.CHANGE, {
-			newDomain
-		});
+		if (configs.dispatchEvent) {
+			this.services.events.dispatchEvent(Events.ZoomDomain.CHANGE, {
+				newDomain
+			});
+		}
 	}
-
 	getZoomRatio() {
 		return Tools.getProperty(
 			this.model.getOptions(),