+++ CHANGELOG.md
@@ -3,6 +3,82 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.39.1](https://github.com/carbon-design-system/carbon-charts/compare/v0.39.0...v0.39.1) (2020-10-05)
+
+
+### Bug Fixes
+
+* **core:** fix division by zero & timescale issues ([#829](https://github.com/carbon-design-system/carbon-charts/issues/829)) ([ce9a2ab](https://github.com/carbon-design-system/carbon-charts/commit/ce9a2abdace0eef5ca6bdffa168992972d7f5848))
+
+
+
+
+
+# [0.39.0](https://github.com/carbon-design-system/carbon-charts/compare/v0.38.8...v0.39.0) (2020-10-01)
+
+
+### Features
+
+* **core:** allow custom labels for total & group in tooltips ([#826](https://github.com/carbon-design-system/carbon-charts/issues/826)) ([6946477](https://github.com/carbon-design-system/carbon-charts/commit/69464778ed4fdd2230cf5ca93162ec11b310e580))
+
+
+
+
+
+## [0.38.8](https://github.com/carbon-design-system/carbon-charts/compare/v0.38.7...v0.38.8) (2020-09-30)
+
+**Note:** Version bump only for package @carbon/charts-monorepo
+
+
+
+
+
+## [0.38.7](https://github.com/carbon-design-system/carbon-charts/compare/v0.38.6...v0.38.7) (2020-09-23)
+
+**Note:** Version bump only for package @carbon/charts-monorepo
+
+
+
+
+
+## [0.38.6](https://github.com/carbon-design-system/carbon-charts/compare/v0.38.5...v0.38.6) (2020-09-22)
+
+**Note:** Version bump only for package @carbon/charts-monorepo
+
+
+
+
+
+## [0.38.5](https://github.com/carbon-design-system/carbon-charts/compare/v0.38.4...v0.38.5) (2020-09-22)
+
+
+### Bug Fixes
+
+* do not dispatch zoom domain change event for all zoom bar behaviors ([#811](https://github.com/carbon-design-system/carbon-charts/issues/811)) ([5fd692f](https://github.com/carbon-design-system/carbon-charts/commit/5fd692f9e71dbf4b3e9052feb6c99cb25156c5da))
+
+
+
+
+
+## [0.38.4](https://github.com/carbon-design-system/carbon-charts/compare/v0.38.3...v0.38.4) (2020-09-21)
+
+
+### Bug Fixes
+
+* **core:** fix valueFormatter not being called on non-axis tooltips ([#809](https://github.com/carbon-design-system/carbon-charts/issues/809)) ([b1ff24c](https://github.com/carbon-design-system/carbon-charts/commit/b1ff24c70c8291ef8dab5585866a12ef99fd6e6f))
+
+
+
+
+
+## [0.38.3](https://github.com/carbon-design-system/carbon-charts/compare/v0.38.2...v0.38.3) (2020-09-21)
+
+**Note:** Version bump only for package @carbon/charts-monorepo
+
+
+
+
+
 ## [0.38.2](https://github.com/carbon-design-system/carbon-charts/compare/v0.38.1...v0.38.2) (2020-09-17)
 
 **Note:** Version bump only for package @carbon/charts-monorepo